Add zimbrasoap (as-is) and patched pysimplesoap
Added current zimbrasoap lib, patched pysimplesoap, and setup.py for
zimbrasoap. This version of pysimplesoap (v1.10) was chosen because it
is the packaged release, but it's missing the "trace" option for the
soap client. Zimbrasoap uses this, meaning the current versions are
incompatible. Committing this way so changes can be tracked.

By v1.10, pysimplesoap switched from using the "trace" option to trace
soap calls to using the Python "logger" module. Later (in git), the
trace option returned, but it now sets logger to log globally.

To prevent zimbrasoap from logging globally when in debug mode, we
should turn tracing on via logger. Possibly should also differentiate
soap trace mode from debug mode (and provide debug output :).