Change debug -> trace, add trace toggle function
pysimplesoap now uses logger for debug output. Utilize logger to print
debug output for only pysimplesoap's client module when tracing is on,
rather than system-wide debug mode in logger that pysimplesoap will be
defaulting to in later versions when using the default trace.

This fixes compatibilty with zimbrasoap and v1.10 of pysimplesoap.